test: consolidate bios_info test into system_collector_test.go
Per PR #75 review: avoid creating a new test file per metric.
